Extract shared XHR helper and fix abort signal listener leak
Move duplicated XHR progress-upload logic from uploadWithProgress
and sideloadWithProgress into a shared xhrUpload() helper. This
fixes the abort signal event listener never being cleaned up after
requests complete, preventing potential memory leaks when
AbortControllers are long-lived.

Also deduplicates createUploadFormData/createSideloadFormData into
a shared buildFormData() with typed wrapper functions.
