+++ packages/media-utils/src/utils/sideload-to-server.ts
@@ -6,109 +6,10 @@ import apiFetch from '@wordpress/api-fetch';
 /**
  * Internal dependencies
  */
-import { buildRestUrl } from './build-rest-url';
 import { createSideloadFormData } from './create-upload-form-data';
 import { transformAttachment } from './transform-attachment';
+import { xhrUpload } from './xhr-upload';
 import type { Attachment, CreateSideloadFile, RestAttachment } from './types';
-
-/**
- * Sideloads a file to the server using XMLHttpRequest with progress tracking.
- *
- * @param file           Media File to Save.
- * @param attachmentId   Parent attachment ID.
- * @param additionalData Additional data to include in the request.
- * @param signal         Abort signal.
- * @param onProgress     Callback for upload progress (0-100).
- * @return Promise resolving to the sideloaded attachment.
- */
-function sideloadWithProgress(
-	file: File,
-	attachmentId: RestAttachment[ 'id' ],
-	additionalData: CreateSideloadFile = {},
-	signal?: AbortSignal,
-	onProgress?: ( progress: number ) => void
-): Promise< Attachment > {
-	return new Promise( ( resolve, reject ) => {
-		// Handle abort signal - check early before creating resources.
-		if ( signal?.aborted ) {
-			reject( new DOMException( 'Aborted', 'AbortError' ) );
-			return;
-		}
-
-		const data = createSideloadFormData( file, additionalData );
-		const xhr = new XMLHttpRequest();
-
-		if ( signal ) {
-			signal.addEventListener( 'abort', () => {
-				xhr.abort();
-			} );
-		}
-
-		// Track upload progress
-		xhr.upload.onprogress = ( event ) => {
-			if ( event.lengthComputable && onProgress ) {
-				const progress = Math.round(
-					( event.loaded / event.total ) * 100
-				);
-				onProgress( progress );
-			}
-		};
-
-		xhr.onload = () => {
-			if ( xhr.status >= 200 && xhr.status < 300 ) {
-				try {
-					const response = JSON.parse(
-						xhr.responseText
-					) as RestAttachment;
-					resolve( transformAttachment( response ) );
-				} catch {
-					reject( new Error( 'Invalid JSON response' ) );
-				}
-			} else {
-				// Try to parse error response
-				try {
-					const errorResponse = JSON.parse( xhr.responseText );
-					reject( errorResponse );
-				} catch {
-					reject(
-						new Error(
-							`Sideload failed with status ${ xhr.status }`
-						)
-					);
-				}
-			}
-		};
-
-		xhr.onerror = () => {
-			reject( new Error( 'Network error during sideload' ) );
-		};
-
-		xhr.onabort = () => {
-			reject( new DOMException( 'Aborted', 'AbortError' ) );
-		};
-
-		// Build the URL using the helper that handles plain permalinks and _locale.
-		const url = buildRestUrl( `/wp/v2/media/${ attachmentId }/sideload` );
-
-		xhr.open( 'POST', url );
-
-		// Set headers
-		xhr.setRequestHeader( 'Accept', 'application/json, */*;q=0.1' );
-
-		// Add nonce header if available
-		if ( apiFetch.nonceMiddleware?.nonce ) {
-			xhr.setRequestHeader(
-				'X-WP-Nonce',
-				apiFetch.nonceMiddleware.nonce
-			);
-		}
-
-		// Include credentials for cookie-based auth
-		xhr.withCredentials = true;
-
-		xhr.send( data );
-	} );
-}
 
 /**
  * Uploads a file to the server without creating an attachment.
@@ -127,24 +28,20 @@ export async function sideloadToServer(
 	additionalData: CreateSideloadFile = {},
 	signal?: AbortSignal,
 	onProgress?: ( progress: number ) => void
-) {
-	// Use XHR when progress tracking is needed, otherwise fall back to apiFetch
+): Promise< Attachment > {
+	const path = `/wp/v2/media/${ attachmentId }/sideload`;
+	const data = createSideloadFormData( file, additionalData );
+
+	// Use XHR when progress tracking is needed, otherwise fall back to apiFetch.
 	if ( onProgress ) {
-		return sideloadWithProgress(
-			file,
-			attachmentId,
-			additionalData,
-			signal,
-			onProgress
+		return transformAttachment(
+			await xhrUpload< RestAttachment >( path, data, signal, onProgress )
 		);
 	}
 
-	// Create upload payload using the shared helper.
-	const data = createSideloadFormData( file, additionalData );
-
 	return transformAttachment(
 		await apiFetch< RestAttachment >( {
-			path: `/wp/v2/media/${ attachmentId }/sideload`,
+			path,
 			body: data,
 			method: 'POST',
 			signal,
